requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
asgiref==3.7.2; python_full_version >= '3.7.0'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
django==4.2.3
django-annoying==0.10.6
django-crontab==0.7.1
django-filter==22.1; python_full_version >= '3.7.0'
djangorestframework==3.14.0; python_version >= '3.6'
graphene==3.2.2
graphene-django==3.1.3
graphene-django-extras==1.0.0
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
idna==3.4; python_version >= '3.5'
pillow==10.0.0
promise==2.3
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
requests==2.31.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
text-unidecode==1.3
typing-extensions==4.7.1; python_version < '3.11'
urllib3==2.0.4; python_full_version >= '3.7.0'
